refactor: domain - repository 아래 이슈 별로 패키지 분리
