+++ android/app/src/main/java/com/woowacourse/ody/data/local/fake/FakeMeetingRepository.kt
@@ -4,7 +4,7 @@ import com.woowacourse.ody.data.remote.ody.entity.meeting.request.MeetingRequest
 import com.woowacourse.ody.data.remote.ody.entity.meeting.response.MeetingResponse
 import com.woowacourse.ody.domain.model.Mate
 import com.woowacourse.ody.domain.model.Meeting
-import com.woowacourse.ody.domain.repository.MeetingRepository
+import com.woowacourse.ody.domain.repository.ody.MeetingRepository
 import java.time.LocalDate
 import java.time.LocalTime
