+++ android/app/src/main/java/com/woowacourse/ody/data/local/fake/FakeNotificationLogRepository.kt
@@ -2,7 +2,7 @@ package com.woowacourse.ody.data.local.fake
 
 import com.woowacourse.ody.domain.model.NotificationLog
 import com.woowacourse.ody.domain.model.NotificationType
-import com.woowacourse.ody.domain.repository.NotificationLogRepository
+import com.woowacourse.ody.domain.repository.ody.NotificationLogRepository
 import java.time.LocalDateTime
 
 object FakeNotificationLogRepository : NotificationLogRepository {
