+++ android/app/src/main/java/com/woowacourse/ody/data/remote/ody/repository/DefaultNotificationLogRepository.kt
@@ -4,7 +4,7 @@ import com.woowacourse.ody.data.remote.ody.entity.toNotificationList
 import com.woowacourse.ody.data.remote.RetrofitClient
 import com.woowacourse.ody.data.remote.ody.service.NotificationService
 import com.woowacourse.ody.domain.model.NotificationLog
-import com.woowacourse.ody.domain.repository.NotificationLogRepository
+import com.woowacourse.ody.domain.repository.ody.NotificationLogRepository
 
 object DefaultNotificationLogRepository : NotificationLogRepository {
     private val notificationService: NotificationService =
