+++ android/app/src/main/java/com/woowacourse/ody/domain/repository/ody/JoinRepository.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.ody.domain.repository
+package com.woowacourse.ody.domain.repository.ody
 
 import com.woowacourse.ody.data.remote.ody.entity.join.request.JoinRequest
 import com.woowacourse.ody.data.remote.ody.entity.meeting.response.MeetingResponse
