+++ android/app/src/main/java/com/woowacourse/ody/presentation/completion/JoinCompleteViewModelFactory.kt
@@ -3,8 +3,8 @@ package com.woowacourse.ody.presentation.completion
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.woowacourse.ody.data.local.db.OdyDatastore
-import com.woowacourse.ody.domain.repository.JoinRepository
-import com.woowacourse.ody.domain.repository.MeetingRepository
+import com.woowacourse.ody.domain.repository.ody.JoinRepository
+import com.woowacourse.ody.domain.repository.ody.MeetingRepository
 
 class JoinCompleteViewModelFactory(
     private val meetingRepository: MeetingRepository,
