+++ android/app/src/main/java/com/woowacourse/ody/presentation/invitecode/InviteCodeViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.map
 import androidx.lifecycle.viewModelScope
-import com.woowacourse.ody.domain.repository.MeetingRepository
+import com.woowacourse.ody.domain.repository.ody.MeetingRepository
 import com.woowacourse.ody.util.Event
 import com.woowacourse.ody.util.emit
 import kotlinx.coroutines.launch
